refactor(linsight): split stepUtils.ts by concern
Pure move, no behaviour change — stepUtils.ts had reached 1055 lines against the
600-line cap, and the narration rules that need editing next sit in the middle
of it.

  execTypes  frame / step / node type contracts
  stepTree   frame merge + flow/timeline grouping
  narration  firstLine + the group narration aside
  activity   tool-call activity buckets
  clarify    user_input parsing / answer composition
  taskStatus status predicates + session pseudo-task split

stepUtils.ts stays as a barrel re-exporting all six, so all 18 call sites are
untouched. The 84 existing tests pass unchanged, and tsc reports the same single
pre-existing error, now at stepTree.ts:443 instead of stepUtils.ts:866.

(cherry picked from commit df50a6034719167f34f1923102d58493936be581)
